Normalize config at load time
