+++ src/types/ConfigSet.ts
@@ -1,4 +1,4 @@
-import Config from './Config';
+import Config from './config/Config';
 
 export default interface ConfigSet {
     [path: string]: Config;
