+++ src/types/config/Config.ts
@@ -1,0 +1,10 @@
+import NormalizedPath from '../NormalizedPath';
+import DependencyRule from './DependencyRule';
+
+export default interface Config {
+    path: NormalizedPath;
+    tags: string[];
+    exports: { [files: string]: string | string[] };
+    dependencies: DependencyRule[];
+    imports: string[];
+};
